refactor : " redis 동적 사용을 server에서 모두 관리하도록 리팩토링 "
