+++ src/gameServer/events/onClose.ts
@@ -1,29 +1,36 @@
-import { getRedisData, getUserIdBySocket, setRedisData } from '../handlers/handlerMethod.js';
+import Server from '../class/server.js';
+import UserSessions from '../class/userSessions.js';
+import { getRoomByUserId, getUserBySocket } from '../handlers/handlerMethod.js';
 import { leaveRoomHandler } from '../handlers/room/leaveRoomHandler.js';
 import { CustomSocket } from '../interface/interface.js';
 import { socketSessions } from '../session/socketSession.js';
 
 export const onClose = (socket: CustomSocket) => async () => {
-  console.log(`게임서버 연결 종료 ${socket.remoteAddress}:${socket.remotePort}`);
+    Server.getInstance().connectingClientCount--;
 
-  const endUserId: number | null = await getUserIdBySocket(socket);
-  if (endUserId) {
-    // 강제종료 했을 경우 방에서 나가기 처리
-    await leaveRoomHandler(socket);
+    console.log(`게임서버 연결 종료 ${socket.remoteAddress}:${socket.remotePort} 연결 중인 클라 ${Server.getInstance().connectingClientCount}`);
 
-    // userData 없애기
-    const liveUserDatas = await getRedisData('userData');
-    console.log('종료 처리 전 liveUserData 수: ', liveUserDatas.length);
-    for (let i = 0; i < liveUserDatas.length; i++) {
-      if (liveUserDatas[i].id === endUserId) {
-        liveUserDatas.splice(i, 1);
-        await setRedisData('userData', liveUserDatas);
-        break;
-      }
+    const closeUser = getUserBySocket(socket);
+    if (!closeUser) {
+        return;
     }
-    console.log('종료 처리 후 liveUserData 수: ', liveUserDatas.length);
 
-    // socketSessions 없애기
-    delete socketSessions[endUserId];
-  }
+    const users = Server.getInstance().getUsers();
+    if (!users) {
+        return;
+    }
+
+    // 참여 중인 방이 있으면 방에서 나감
+    const leaveRoom = getRoomByUserId(closeUser.getId());
+    if (leaveRoom) {
+        await leaveRoomHandler(socket);
+    }
+
+    // users에서 삭제
+    const remainUsers = users.filter((user: UserSessions) => user.getId() !== closeUser.getId());
+    Server.getInstance().setUsers(remainUsers);
+
+    console.log(`로그인 중인 클라 ${Server.getInstance().getUsers().length}`);
+
+    delete socketSessions[closeUser.getId()];
 };
