+++ src/gameServer/handlers/card/useCardHandler.ts
@@ -2,24 +2,23 @@ import { config } from '../../../config/config.js';
 import {
   CustomSocket,
   UseCardRequest,
-  UseCardResponse,
-  Room,
-  User,
   CharacterPositionData,
   skillCardDBData,
   position
 } from '../../../gameServer/interface/interface.js';
-import { CardType, GlobalFailCode, RoleType } from '../enumTyps.js';
+import { CardType, RoleType } from '../enumTyps.js';
 import { sendPacket } from '../../../packet/createPacket.js';
-import { getRedisData, getUserIdBySocket, setRedisData } from '../handlerMethod.js';
-import { userUpdateNotification } from '../notification/userUpdate.js';
+import { getUserBySocket } from '../handlerMethod.js';
 import { socketSessions } from '../../session/socketSession.js';
 import { monsterAiDatas } from '../coreMethod/monsterMove.js';
-import { monsterReward, setCardRewards, setStatRewards } from '../coreMethod/monsterReward.js';
 import { gameEndNotification } from '../notification/gameEnd.js';
 import Server from '../../class/server.js';
 import { shoppingUserIdSessions } from '../../session/shoppingSession.js';
 import { dbManager } from '../../../database/user/user.db.js';
+import GameRoom from '../../class/room.js';
+import UserSessions from '../../class/userSessions.js';
+import { userUpdateNotification } from '../notification/userUpdate.js';
+import { monsterReward, setCardRewards, setStatRewards } from '../coreMethod/monsterReward.js';
 
 const characterBuffStatus: { [characterId: number]: number[] } = {};
 
@@ -37,70 +36,53 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
   const { cardType, targetUserId: targetUserIdRaw } = payload as UseCardRequest;
 
   const targetUserId = Number(targetUserIdRaw);
-  let responseData: UseCardResponse = {
-    success: true,
-    failCode: GlobalFailCode.NONE
-  };
   console.log('cardType: ', cardType);
   try {
     // redisUser 정보 찾기
-    const userId: number | null = await getUserIdBySocket(socket);
-    if (userId === null) {
+    const user = getUserBySocket(socket);
+    if (!user) {
       console.error('카드 사용자의 정보를 찾을 수 없습니다.');
       return;
     }
 
     // room 정보 찾기
-    const rooms: Room[] | undefined = await getRedisData('roomData');
-    if (!rooms || rooms.length === 0) {
-      console.error('서버에 Rooms정보가 존재하지 않습니다.');
-      return;
-    }
-    let room: Room | undefined;
-    for (let i = 0; i < rooms.length; i++) {
-      for (let j = 0; j < rooms[i].users.length; j++) {
-        if (rooms[i].users[j].id === userId) {
-          room = rooms[i];
-        }
-      }
-    }
-    if (!room || !room.users) {
+    const room: GameRoom | undefined = Server.getInstance()
+      .getRooms()
+      .find((room: GameRoom) => room.getUsers().some((roomUser: UserSessions) => roomUser.getId() === user.getId()));
+    if (!room) {
       console.error('카드 사용자가 속한 room 정보를 찾을 수 없습니다.');
       return;
     }
 
-    // room 의 user정보 찾기
-    let user: User | null = null;
-    for (let i = 0; i < room.users.length; i++) {
-      if (room.users[i].id === userId) {
-        user = room.users[i];
-        break;
-      }
-    }
-    if (!user || !user.character) return;
-    if (user.character.aliveState === false || user.character.hp <= 0) {
+    if (!user || !user.getCharacter()) return;
+    if (user.getCharacter().aliveState === false || user.getCharacter().hp <= 0) {
       console.log(
-        `죽어있는 대상은 행동을 할 수 없습니다. 상태: ${user.character.aliveState}, 체력: ${user.character.hp}`
+        `죽어있는 대상은 행동을 할 수 없습니다. 상태: ${user.getCharacter().aliveState}, 체력: ${user.getCharacter().hp}`
       );
       return;
     }
 
     // target 정보 찾기
-    let target: User | null = null;
-    for (let i = 0; i < room.users.length; i++) {
-      if (room.users[i].id === targetUserId) {
-        target = room.users[i];
+    let target: UserSessions | null = null;
+    for (let i = 0; i < room.getUsers().length; i++) {
+      if (room.getUsers()[i].getId() === targetUserId) {
+        target = room.getUsers()[i];
       }
     }
 
-    if (user.character.roleType === RoleType.SUR5VAL) {
-      // 캐릭터 위치 정보 찾기
-      const characterPositions: { [roomId: number]: CharacterPositionData[] } =
-        await getRedisData('characterPositionDatas');
+    // 캐릭터 위치 정보 찾기
+    if (user.getCharacter().roleType === RoleType.SUR5VAL) {
+      const characterPositions = Server.getInstance()
+        .getPositions()
+        .find((PositionSessions) => PositionSessions.getPositionRoomId() === room.getRoomId());
+      if (!characterPositions) {
+        console.error('characterPositions를 찾지 못하였습니다.');
+        return;
+      }
       // 공격 가능 여부 확인
-      for (let i = 0; i < characterPositions[room.id].length; i++) {
-        let userPos: CharacterPositionData = characterPositions[room.id][i];
-        if (user !== null && userPos.id === user.id) {
+      for (let i = 0; i < characterPositions.getCharacterPositions().length; i++) {
+        let userPos: CharacterPositionData = characterPositions.getCharacterPositions()[i];
+        if (user !== null && userPos.id === user.getId()) {
           if (
             // 공격자가 실외에 위치해 있는 경우
             !(-23 <= userPos.x && userPos.x <= -12 && 5 <= userPos.y && userPos.y <= 10) && // 건물 1
@@ -118,9 +100,9 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
             !(-2 <= userPos.x && userPos.x <= -1 && -8.5 <= userPos.y && userPos.y <= -7.5) && // 부쉬 5
             !(4 <= userPos.x && userPos.x <= 5 && -8.5 <= userPos.y && userPos.y <= -7.5) // 부쉬 6
           ) {
-            for (let j = 0; j < characterPositions[room.id].length; j++) {
-              let characterPos: CharacterPositionData = characterPositions[room.id][j];
-              if (target !== null && characterPos.id === target.id) {
+            for (let j = 0; j < characterPositions.getCharacterPositions().length; j++) {
+              let characterPos: CharacterPositionData = characterPositions.getCharacterPositions()[j];
+              if (target !== null && characterPos.id === target.getId()) {
                 if (
                   (-23 <= characterPos.x && characterPos.x <= -12 && 5 <= characterPos.y && characterPos.y <= 10) || // 건물 1
                   (-7 <= characterPos.x && characterPos.x <= 0 && 5 <= characterPos.y && characterPos.y <= 10) || // 건물 2
@@ -165,15 +147,14 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.SUR5VER_BASIC_SKILL, user, target, 0))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
 
         // 공격 실행
-        await attackTarget(user.id, room.id, 1, target.id);
+        attackTarget(user.getId(), room.getRoomId(), 1, target.getId());
         sendAnimation(room, target, 1);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 5;
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 5;
             break;
           }
         }
@@ -188,16 +169,15 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.MAGICIAN_BASIC_SKILL, user, target, 2))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 2;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 2;
 
         // 스킬 실행1
-        await attackTarget(user.id, room.id, 1, target.id);
+        attackTarget(user.getId(), room.getRoomId(), 1, target.getId());
         let index: number | null = null;
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 5;
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 5;
             index = i;
             break;
           }
@@ -206,8 +186,8 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
 
         // 스킬 실행2
         setTimeout(async () => {
-          await attackTarget(user.id, room.id, 1.5, target.id);
-          if (index) monsterAiDatas[room.id][index].animationDelay = 5;
+          attackTarget(user.getId(), room.getRoomId(), 1.5, target.getId());
+          if (index) monsterAiDatas[room.getRoomId()][index].animationDelay = 5;
           sendAnimation(room, target, 3);
         }, 800);
 
@@ -221,15 +201,14 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.ARCHER_BASIC_SKILL, user, target, 2))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 2;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 2;
 
         // 공격 실행
-        await attackTarget(user.id, room.id, 2, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 4;
+        attackTarget(user.getId(), room.getRoomId(), 2, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 4;
             break;
           }
         }
@@ -244,15 +223,14 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.ROGUE_BASIC_SKILL, user, target, 3))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 3;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 3;
 
         // 공격 실행
-        await attackTarget(user.id, room.id, 2.5, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 6;
+        attackTarget(user.getId(), room.getRoomId(), 2.5, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 6;
             break;
           }
         }
@@ -264,13 +242,13 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 설명: MP소모: 2, 투사의 의지로 방패는 견고해지고 검은 강력해집니다.
       case CardType.WARRIOR_BASIC_SKILL:
         // 공격 유효성 검증
-        if (user.character.mp < 2) {
+        if (user.getCharacter().mp < 2) {
           console.log('마나가 부족합니다.');
           return;
         }
 
         // 버프 스킬 실행
-        if (await changeStatus(2, user, rooms, room, 0, 2, 2, 15000, CardType.WARRIOR_BASIC_SKILL)) {
+        if (changeStatus(2, user, room, 0, 2, 2, 15000, CardType.WARRIOR_BASIC_SKILL)) {
           sendAnimation(room, user, 6);
         }
 
@@ -284,16 +262,15 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.MAGICIAN_EXTENDED_SKILL, user, target, 3))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 3;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 3;
 
         // 스킬 실행1
-        await attackTarget(user.id, room.id, 1.2, target.id);
+        attackTarget(user.getId(), room.getRoomId(), 1.2, target.getId());
         let index: number | null = null;
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 4;
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 4;
             index = i;
             break;
           }
@@ -302,15 +279,15 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
 
         // 스킬 실행2
         setTimeout(async () => {
-          await attackTarget(user.id, room.id, 1.4, target.id);
-          if (index) monsterAiDatas[room.id][index].animationDelay = 4;
+          attackTarget(user.getId(), room.getRoomId(), 1.4, target.getId());
+          if (index) monsterAiDatas[room.getRoomId()][index].animationDelay = 4;
           sendAnimation(room, target, 1);
         }, 600);
 
         // 스킬 실행3
         setTimeout(async () => {
-          await attackTarget(user.id, room.id, 1.7, target.id);
-          if (index) monsterAiDatas[room.id][index].animationDelay = 4;
+          attackTarget(user.getId(), room.getRoomId(), 1.7, target.getId());
+          if (index) monsterAiDatas[room.getRoomId()][index].animationDelay = 4;
           sendAnimation(room, target, 1);
         }, 1200);
 
@@ -321,13 +298,13 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 설명: Mp소모: 4, 강력한 회오리 바람의 정령을 소환합니다. 정령은 캐릭터를 따라다니며 주변 적을 지속적으로 공격합니다.
       case CardType.ARCHER_EXTENDED_SKILL:
         // 공격 유효성 검증
-        if (user.character.mp < 4) {
+        if (user.getCharacter().mp < 4) {
           console.log('마나가 부족합니다.');
           return;
         }
 
         // 정령 버프 실행
-        if (await summonSpiritBuff(4, user, room, rooms, 0.8, 5, 2, 15000, 4000, CardType.ARCHER_EXTENDED_SKILL)) {
+        if (await summonSpiritBuff(4, user, room, 0.8, 5, 2, 15000, 4000, CardType.ARCHER_EXTENDED_SKILL)) {
           sendAnimation(room, user, 6);
         }
 
@@ -337,13 +314,13 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 설명: Mp소모: 4, 날카로운 그림자 칼날의 정령을 소환합니다. 정령은 캐릭터를 따라다니며 주변 적을 지속적으로 공격합니다.
       case CardType.ROGUE_EXTENDED_SKILL:
         // 공격 유효성 검증
-        if (user.character.mp < 4) {
+        if (user.getCharacter().mp < 4) {
           console.log('마나가 부족합니다.');
           return;
         }
 
         // 정령 버프 실행
-        if (await summonSpiritBuff(4, user, room, rooms, 1.2, 5, 1, 15000, 4000, CardType.ROGUE_EXTENDED_SKILL)) {
+        if (await summonSpiritBuff(4, user, room, 1.2, 5, 1, 15000, 4000, CardType.ROGUE_EXTENDED_SKILL)) {
           sendAnimation(room, user, 6);
         }
 
@@ -357,15 +334,14 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.WARRIOR_EXTENDED_SKILL, user, target, 3))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 3;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 3;
 
         // 공격 실행
-        await attackTargetTrueDamage(userId, room.id, 3, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 6;
+        attackTargetTrueDamage(user.getId(), room.getRoomId(), 3, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 6;
             break;
           }
         }
@@ -380,15 +356,14 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.MAGICIAN_FINAL_SKILL, user, target, 4))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 4;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 4;
 
         // 스킬 실행
-        await attackTarget(user.id, room.id, 6, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 6;
+        attackTarget(user.getId(), room.getRoomId(), 6, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 6;
             break;
           }
         }
@@ -400,13 +375,13 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 설명: Mp소모: 5, 전장의 수호자로서 모든 아군에게 일정시간동안 유지되는 불굴의 힘과 방어력을 부여합니다.
       case CardType.ARCHER_FINAL_SKILL:
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
+        user.getCharacter().coolDown = Date.now();
 
         // 버프 실행
-        await partyBuff(5, user, rooms, room, 0, 3, 3, 15000, CardType.WARRIOR_FINAL_SKILL);
-        for (let i = 0; i < room.users.length; i++) {
-          if (room.users[i].character.roleType === RoleType.SUR5VAL) {
-            sendAnimation(room, room.users[i], 6);
+        partyBuff(5, user, room, 0, 3, 3, 15000, CardType.WARRIOR_FINAL_SKILL);
+        for (let i = 0; i < room.getUsers().length; i++) {
+          if (room.getUsers()[i].getCharacter().roleType === RoleType.SUR5VAL) {
+            sendAnimation(room, room.getUsers()[i], 6);
           }
         }
         break;
@@ -419,16 +394,15 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.ROGUE_FINAL_SKILL, user, target, 5))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 5;
-        await setRedisData('roomData', rooms);
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 5;
 
         // 스킬 실행
-        await changeStatus(0, target, rooms, room, 0, 0, -user.character.attack, 10000, CardType.ROGUE_FINAL_SKILL);
-        await attackTarget(user.id, room.id, 4, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 6;
+        changeStatus(0, target, room, 0, 0, -user.getCharacter().attack, 10000, CardType.ROGUE_FINAL_SKILL);
+        attackTarget(user.getId(), room.getRoomId(), 4, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 6;
             break;
           }
         }
@@ -444,16 +418,18 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!(await attackPossible(CardType.WARRIOR_FINAL_SKILL, user, target, 4))) return;
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 4;
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 4;
 
         // 스킬 실행
-        user.character.hp = Math.min(user.character.hp + user.character.attack * 2, user.character.maxHp);
-        await setRedisData('roomData', rooms);
-        await attackTarget(user.id, room.id, 3, target.id);
-        for (let i = 0; i < monsterAiDatas[room.id].length; i++) {
-          if (monsterAiDatas[room.id][i].id === target.id) {
-            monsterAiDatas[room.id][i].animationDelay = 6;
+        user.getCharacter().hp = Math.min(
+          user.getCharacter().hp + user.getCharacter().attack * 2,
+          user.getCharacter().maxHp
+        );
+        attackTarget(user.getId(), room.getRoomId(), 3, target.getId());
+        for (let i = 0; i < monsterAiDatas[room.getRoomId()].length; i++) {
+          if (monsterAiDatas[room.getRoomId()][i].id === target.getId()) {
+            monsterAiDatas[room.getRoomId()][i].animationDelay = 6;
             break;
           }
         }
@@ -464,7 +440,7 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 이름: 보스 기본 공격
       // 설명: 본인 주위의 적들을 공격합니다.
       case CardType.BOSS_BASIC_SKILL:
-        await attackRagne(user, user, rooms, room, 1, 5, 3);
+        attackRagne(user, user, room, 1, 5, 3);
         break;
 
       // 이름: 대재앙의 강타
@@ -477,30 +453,30 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         if (!initGameInfo) return;
 
         const characterAttackCool = initGameInfo.coolTime;
-        if (Date.now() - user.character.coolDown < characterAttackCool) {
+        if (Date.now() - user.getCharacter().coolDown < characterAttackCool) {
           console.log('공격 쿨타임 중입니다.');
           return;
         }
-        if (user.character.mp < 10) {
+        if (user.getCharacter().mp < 10) {
           console.log('마나가 부족합니다.');
           return;
         }
-        if (target.character.roleType !== RoleType.SUR5VAL) {
+        if (target.getCharacter().roleType !== RoleType.SUR5VAL) {
           console.log('아군을 공격할 수는 없습니다.');
           return;
         }
-        if (target.character.hp <= 0) {
+        if (target.getCharacter().hp <= 0) {
           console.log('살아있는 적만 공격할 수 있습니다.');
           return;
         }
 
         // 공격 자원 처리
-        user.character.coolDown = Date.now();
-        user.character.mp -= 10;
+        user.getCharacter().coolDown = Date.now();
+        user.getCharacter().mp -= 10;
 
         // 스킬 실행
-        let isRange = await attackRagne(user, target, rooms, room, 0.6, 4, 3);
-        if (isRange) await attackTarget(user.id, room.id, 0.8, target.id);
+        let isRange = attackRagne(user, target, room, 0.6, 4, 3);
+        if (isRange) attackTarget(user.getId(), room.getRoomId(), 0.8, target.getId());
         sendAnimation(room, target, 11);
         break;
 
@@ -515,49 +491,49 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 이름: 순수한 이슬
       // 설명: 맑고 순수한 이슬 한 방울이 체력을 20 회복해준다! 지친 몸에 활력을 더해 다시 전투에 나설 수 있도록 돕는다.
       case CardType.BASIC_HP_POTION:
-        await usePotion(user, rooms, room, CardType.BASIC_HP_POTION);
+        usePotion(user, room, CardType.BASIC_HP_POTION);
         break;
 
       // 이름: 마력의 이슬
       // 설명: 맑고 순수한 이슬 한 방울이 마력을 4 회복해준다! 고갈된 마법 에너지를 되살려 새로운 주문을 준비하자.
       case CardType.BASIC_MP_POTION:
-        await usePotion(user, rooms, room, CardType.BASIC_MP_POTION);
+        usePotion(user, room, CardType.BASIC_MP_POTION);
         break;
 
       // 이름: 치유의 빛
       // 설명: 은은한 치유의 빛이 체력을 30 회복해준다! 깊은 상처를 어루만지고 전투의 피로를 씻어내는 신비한 물약.
       case CardType.ADVANCED_HP_POTION:
-        await usePotion(user, rooms, room, CardType.ADVANCED_HP_POTION);
+        usePotion(user, room, CardType.ADVANCED_HP_POTION);
         break;
 
       // 이름: 마력의 빛
       // 설명: 은은한 마나의 빛이 마력을 6 회복해준다! 흐릿했던 마법의 기운을 선명하게 채워주는 신비한 물약.
       case CardType.ADVANCED_MP_POTION:
-        await usePotion(user, rooms, room, CardType.ADVANCED_MP_POTION);
+        usePotion(user, room, CardType.ADVANCED_MP_POTION);
         break;
 
       // 이름: 생명의 숨결
       // 설명: 신비로운 생명의 기운이 체력을 50 회복해준다! 생명의 근원이 담긴 이 물약은 가장 극한의 상황에서도 새로운 힘을 불어넣는다.
       case CardType.MASTER_HP_POTION:
-        await usePotion(user, rooms, room, CardType.MASTER_HP_POTION);
+        usePotion(user, room, CardType.MASTER_HP_POTION);
         break;
 
       // 이름: 마력의 숨결
       // 설명: 신비로운 마력의 기운이 마력을 10 회복해준다! 극한의 상황에서도 강력한 주문을 사용할 수 있는 힘을 불어넣는다.
       case CardType.MASTER_MP_POTION:
-        await usePotion(user, rooms, room, CardType.MASTER_MP_POTION);
+        usePotion(user, room, CardType.MASTER_MP_POTION);
         break;
 
       // 이름: 성장의 작은 불꽃
       // 설명: 작은 불꽃이 당신의 성장을 돕습니다. 경험치 +20
       case CardType.BASIC_EXP_POTION:
-        await usePotion(user, rooms, room, CardType.BASIC_EXP_POTION);
+        usePotion(user, room, CardType.BASIC_EXP_POTION);
         break;
 
       // 이름: 무한 성장의 불길
       // 설명: 끝없는 불길로 압도적인 성장을 경험하세요. 경험치 +50
       case CardType.MASTER_EXP_POTION:
-        await usePotion(user, rooms, room, CardType.MASTER_EXP_POTION);
+        usePotion(user, room, CardType.MASTER_EXP_POTION);
         break;
 
       // 이름: 용기의 정수
@@ -565,9 +541,9 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       case CardType.ATTACK_POTION: {
         // 해당 아이템 보유여부 검사 및 개수 차감
         let isOwned: boolean = false;
-        for (let i = 0; i < user.character.handCards.length; i++) {
-          if (user.character.handCards[i].type === cardType && user.character.handCards[i].count > 0) {
-            user.character.handCards[i].count--;
+        for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+          if (user.getCharacter().handCards[i].type === cardType && user.getCharacter().handCards[i].count > 0) {
+            user.getCharacter().handCards[i].count--;
             isOwned = true;
             break;
           }
@@ -578,7 +554,7 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         }
 
         // 버프 실행
-        await changeStatus(0, user, rooms, room, 0, 0, 5, 30000, CardType.ATTACK_POTION);
+        changeStatus(0, user, room, 0, 0, 5, 30000, CardType.ATTACK_POTION);
         break;
       }
 
@@ -587,9 +563,9 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       case CardType.DEFENSE_PORTION:
         // 해당 아이템 보유여부 검사 및 개수 차감
         let isOwned: boolean = false;
-        for (let i = 0; i < user.character.handCards.length; i++) {
-          if (user.character.handCards[i].type === cardType && user.character.handCards[i].count > 0) {
-            user.character.handCards[i].count--;
+        for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+          if (user.getCharacter().handCards[i].type === cardType && user.getCharacter().handCards[i].count > 0) {
+            user.getCharacter().handCards[i].count--;
             isOwned = true;
             break;
           }
@@ -600,7 +576,7 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
         }
 
         // 버프 실행
-        await changeStatus(0, user, rooms, room, 0, 6, 0, 30000, CardType.DEFENSE_PORTION);
+        changeStatus(0, user, room, 0, 6, 0, 30000, CardType.DEFENSE_PORTION);
         break;
 
       // 장비 301 ~ 400 // 장비 301 ~ 400  / 장비 301 ~ 400 // 장비 301 ~ 400 // 장비 301 ~ 400 // 장비 301 ~ 400 // 장비 301 ~ 400 // 장비 301 ~ 400 //
@@ -608,91 +584,91 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
       // 이름: 탐험가의 무기
       // 설명: 야생에서 빛을 발하는 무기. 공격력+3
       case CardType.EXPLORER_WEAPON:
-        await equipWeapon(user, rooms, room, CardType.EXPLORER_WEAPON);
+        equipWeapon(user, room, CardType.EXPLORER_WEAPON);
         break;
 
       // 이름: 탐험가의 투구
       // 설명: 거친 환경을 견디는 투구. 체력+21
       case CardType.EXPLORER_HEAD:
-        await equipItem(user, rooms, room, 0, CardType.EXPLORER_HEAD);
+        equipItem(user, room, 0, CardType.EXPLORER_HEAD);
         break;
 
       // 이름: 탐험가의 갑옷
       // 설명: 유연성과 보호력을 겸비한 갑옷. 방어력+1, 체력+17
       case CardType.EXPLORER_ARMOR:
-        await equipItem(user, rooms, room, 1, CardType.EXPLORER_ARMOR);
+        equipItem(user, room, 1, CardType.EXPLORER_ARMOR);
         break;
 
       // 이름: 탐험가의 망토
       // 설명: 바람과 비를 막아주는 견고한 망토. 방어력+2, 체력+13
       case CardType.EXPLORER_CLOAK:
-        await equipItem(user, rooms, room, 2, CardType.EXPLORER_CLOAK);
+        equipItem(user, room, 2, CardType.EXPLORER_CLOAK);
         break;
 
       // 이름: 탐험가의 장갑
       // 설명: 정밀한 손놀림을 돕는 장갑. 공격력+1, 방어력+1, 체력+10
       case CardType.EXPLORER_GLOVE:
-        await equipItem(user, rooms, room, 3, CardType.EXPLORER_GLOVE);
+        equipItem(user, room, 3, CardType.EXPLORER_GLOVE);
         break;
 
       // 이름: 영웅의 무기
       // 설명: 강력한 적도 제압 가능한 놀라운 무기. 공격력+6
       case CardType.HERO_WEAPON:
-        await equipWeapon(user, rooms, room, CardType.HERO_WEAPON);
+        equipWeapon(user, room, CardType.HERO_WEAPON);
         break;
 
       // 이름: 영웅의 투구
       // 설명: 영광스러운 전투를 상징하는 투구. 체력+30
       case CardType.HERO_HEAD:
-        await equipItem(user, rooms, room, 0, CardType.HERO_HEAD);
+        equipItem(user, room, 0, CardType.HERO_HEAD);
         break;
 
       // 이름: 영웅의 갑옷
       // 설명: 방어와 위엄을 겸비한 갑옷. 방어력+2, 체력+22
       case CardType.HERO_ARMOR:
-        await equipItem(user, rooms, room, 1, CardType.HERO_ARMOR);
+        equipItem(user, room, 1, CardType.HERO_ARMOR);
         break;
 
       // 이름: 영웅의 망토
       // 설명: 마법의 힘이 깃든 망토. 방어력+4, 체력+14
       case CardType.HERO_CLOAK:
-        await equipItem(user, rooms, room, 2, CardType.HERO_CLOAK);
+        equipItem(user, room, 2, CardType.HERO_CLOAK);
         break;
 
       // 이름: 영웅의 장갑
       // 설명: 전투 기술을 극대화하는 장갑. 공격력+2, 방어력+2, 체력+12
       case CardType.HERO_GLOVE:
-        await equipItem(user, rooms, room, 3, CardType.HERO_GLOVE);
+        equipItem(user, room, 3, CardType.HERO_GLOVE);
         break;
 
       // 이름: 전설의 무기
       // 설명: 신화적 힘이 담긴 무기, 적에게 파멸을 선사한다. 공격력+10
       case CardType.LEGENDARY_WEAPON:
-        await equipWeapon(user, rooms, room, CardType.LEGENDARY_WEAPON);
+        equipWeapon(user, room, CardType.LEGENDARY_WEAPON);
         break;
 
       // 이름: 전설의 투구
       // 설명: 신성한 보호막을 제공하는 투구, 사용자를 불멸로 이끈다. 체력+40
       case CardType.LEGENDARY_HEAD:
-        await equipItem(user, rooms, room, 0, CardType.LEGENDARY_HEAD);
+        equipItem(user, room, 0, CardType.LEGENDARY_HEAD);
         break;
 
       // 이름: 전설의 갑옷
       // 설명: 황금빛 문양과 마법이 깃든 최강의 갑옷. 방어력+3, 체력+28
       case CardType.LEGENDARY_ARMOR:
-        await equipItem(user, rooms, room, 1, CardType.LEGENDARY_ARMOR);
+        equipItem(user, room, 1, CardType.LEGENDARY_ARMOR);
         break;
 
       // 이름: 전설의 망토
       // 설명: 우주의 에너지가 깃든 최강의 망토. 방어력+6, 체력+16
       case CardType.LEGENDARY_CLOAK:
-        await equipItem(user, rooms, room, 2, CardType.LEGENDARY_CLOAK);
+        equipItem(user, room, 2, CardType.LEGENDARY_CLOAK);
         break;
 
       // 이름: 전설의 장갑
       // 설명: 신화의 기술과 힘을 담은 승리의 장갑. 공격력+3, 방어력+3, 체력+16
       case CardType.LEGENDARY_GLOVE:
-        await equipItem(user, rooms, room, 3, CardType.LEGENDARY_GLOVE);
+        equipItem(user, room, 3, CardType.LEGENDARY_GLOVE);
         break;
     }
   } catch (error) {
@@ -702,32 +678,32 @@ export const useCardHandler = async (socket: CustomSocket, payload: Object): Pro
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 범위 공격
-const attackRagne = async (
-  attacker: User,
-  center: User,
-  rooms: Room[],
-  room: Room,
+const attackRagne = (
+  attacker: UserSessions,
+  center: UserSessions,
+  room: GameRoom,
   skillCoeffcient: number,
   range: number,
   maxTarget: number
-): Promise<Boolean> => {
+) => {
   // 살아있는 플레이어 추출
-  const alivePlayers: [User, number][] = [];
-  for (let i = 0; i < room.users.length; i++) {
-    if (room.users[i].character.roleType === RoleType.SUR5VAL && room.users[i].character.hp > 0) {
-      alivePlayers.push([room.users[i], 0]);
+  const alivePlayers: [UserSessions, number][] = [];
+  for (let i = 0; i < room.getUsers().length; i++) {
+    if (room.getUsers()[i].getCharacter().roleType === RoleType.SUR5VAL && room.getUsers()[i].getCharacter().hp > 0) {
+      alivePlayers.push([room.getUsers()[i], 0]);
     }
   }
 
   // 중심 위치 조회
-  const characterPositions: { [roomId: number]: CharacterPositionData[] | undefined } | undefined =
-    await getRedisData('characterPositionDatas');
+  const characterPositions = Server.getInstance()
+    .getPositions()
+    .find((PositionSessions) => PositionSessions.getPositionRoomId() === room.getRoomId());
   if (characterPositions === undefined) return false;
-  const positionDatas: CharacterPositionData[] | undefined = characterPositions[room.id];
+  const positionDatas: CharacterPositionData[] | undefined = characterPositions.getCharacterPositions();
   if (positionDatas === undefined) return false;
   let bossPosition: position | null = null;
   for (let i = 0; i < positionDatas.length; i++) {
-    if (center.id === positionDatas[i].id) {
+    if (center.getId() === positionDatas[i].id) {
       bossPosition = { x: positionDatas[i].x, y: positionDatas[i].y };
       break;
     }
@@ -737,7 +713,7 @@ const attackRagne = async (
   // 보스와의 거리가 너무 먼 alivePlayer를 splice 처리
   for (let i = 0; i < alivePlayers.length; i++) {
     for (let j = 0; j < positionDatas.length; j++) {
-      if (alivePlayers[i][0].id === positionDatas[j].id) {
+      if (alivePlayers[i][0].getId() === positionDatas[j].id) {
         const distance = (bossPosition.x - positionDatas[j].x) ** 2 + (bossPosition.y - positionDatas[j].y) ** 2;
         if (distance > range ** 2) {
           alivePlayers.splice(i, 1);
@@ -760,11 +736,9 @@ const attackRagne = async (
   let isAttack: boolean = false;
   for (let i = 0; i < alivePlayers.length; i++) {
     // 공격 가능 여부 확인
-    for (let j = 0; j < (characterPositions[room.id] as CharacterPositionData[]).length; j++) {
-      let attackerPos: CharacterPositionData = (characterPositions[room.id] as CharacterPositionData[])[j];
-      console.log('attacker.id: ', attacker.id);
-      console.log('attackerPos.id: ', attackerPos.id);
-      if (attacker !== null && attackerPos.id === attacker.id) {
+    for (let j = 0; j < characterPositions.getCharacterPositions().length; j++) {
+      let attackerPos: CharacterPositionData = characterPositions.getCharacterPositions()[j];
+      if (attacker !== null && attackerPos.id === attacker.getId()) {
         if (
           // 공격자가 실외에 위치해 있는 경우
           !(-23 <= attackerPos.x && attackerPos.x <= -12 && 5 <= attackerPos.y && attackerPos.y <= 10) && // 건물 1
@@ -782,11 +756,9 @@ const attackRagne = async (
           !(-2 <= attackerPos.x && attackerPos.x <= -1 && -8.5 <= attackerPos.y && attackerPos.y <= -7.5) && // 부쉬 5
           !(4 <= attackerPos.x && attackerPos.x <= 5 && -8.5 <= attackerPos.y && attackerPos.y <= -7.5) // 부쉬 6
         ) {
-          for (let k = 0; k < (characterPositions[room.id] as CharacterPositionData[]).length; k++) {
-            let characterPos: CharacterPositionData = (characterPositions[room.id] as CharacterPositionData[])[k];
-            console.log('alivePlayers[i][0].id: ', alivePlayers[i][0].id);
-            console.log('characterPos.id: ', characterPos.id);
-            if (alivePlayers[i][0] !== null && characterPos.id === alivePlayers[i][0].id) {
+          for (let k = 0; k < characterPositions.getCharacterPositions().length; k++) {
+            let characterPos: CharacterPositionData = characterPositions.getCharacterPositions()[k];
+            if (alivePlayers[i][0] !== null && characterPos.id === alivePlayers[i][0].getId()) {
               if (
                 (-23 <= characterPos.x && characterPos.x <= -12 && 5 <= characterPos.y && characterPos.y <= 10) || // 건물 1
                 (-7 <= characterPos.x && characterPos.x <= 0 && 5 <= characterPos.y && characterPos.y <= 10) || // 건물 2
@@ -821,15 +793,18 @@ const attackRagne = async (
     }
 
     if (isAttack) {
-      const damage = Math.max(attacker.character.attack * skillCoeffcient - alivePlayers[i][0].character.armor, 0);
-      alivePlayers[i][0].character.hp -= Math.round(damage);
-      if (alivePlayers[i][0].character.hp <= 0) {
-        alivePlayers[i][0].character.aliveState = false;
-        alivePlayers[i][0].character.stateInfo.state = 15;
-        alivePlayers[i][0].character.hp = 0;
-        for (let i = 0; i < shoppingUserIdSessions[room.id].length; i++) {
-          if (shoppingUserIdSessions[room.id][i][0] === attacker.id) {
-            shoppingUserIdSessions[room.id].splice(i, 1);
+      const damage = Math.max(
+        attacker.getCharacter().attack * skillCoeffcient - alivePlayers[i][0].getCharacter().armor,
+        0
+      );
+      alivePlayers[i][0].getCharacter().hp -= Math.round(damage);
+      if (alivePlayers[i][0].getCharacter().hp <= 0) {
+        alivePlayers[i][0].getCharacter().aliveState = false;
+        alivePlayers[i][0].getCharacter().stateInfo.state = 15;
+        alivePlayers[i][0].getCharacter().hp = 0;
+        for (let i = 0; i < shoppingUserIdSessions[room.getRoomId()].length; i++) {
+          if (shoppingUserIdSessions[room.getRoomId()][i][0] === attacker.getId()) {
+            shoppingUserIdSessions[room.getRoomId()].splice(i, 1);
             break;
           }
         }
@@ -838,28 +813,27 @@ const attackRagne = async (
     }
   }
 
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
+  userUpdateNotification(room);
 
   return true;
 };
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 타겟이 된 적 공격
-const attackTarget = async (attackerId: number, roomId: number, skillCoeffcient: number, targetId: number) => {
+const attackTarget = (attackerId: number, roomId: number, skillCoeffcient: number, targetId: number) => {
   // 공격 실행 중 라운드가 바뀌지 않았는지 검사
-  const nowRooms = await getRedisData('roomData');
+  const nowRooms = Server.getInstance().getRooms();
   let isChanged: boolean = true;
-  let nowRoom: Room | null = null;
+  let nowRoom: GameRoom | null = null;
   for (let i = 0; i < nowRooms.length; i++) {
-    if (nowRooms[i].id === roomId) {
+    if (nowRooms[i].getRoomId() === roomId) {
       nowRoom = nowRooms[i];
       break;
     }
   }
   if (!nowRoom) return;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === targetId) isChanged = false;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === targetId) isChanged = false;
   }
   if (isChanged) {
     console.log('공격 실행 중 라운드 바뀌어서 몬스터 사라짐');
@@ -867,10 +841,10 @@ const attackTarget = async (attackerId: number, roomId: number, skillCoeffcient:
   }
 
   // 공격자의 정보 가져오기
-  let attacker: User | null = null;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === attackerId) {
-      attacker = nowRoom.users[i];
+  let attacker: UserSessions | null = null;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === attackerId) {
+      attacker = nowRoom.getUsers()[i];
       break;
     }
   }
@@ -880,10 +854,10 @@ const attackTarget = async (attackerId: number, roomId: number, skillCoeffcient:
   }
 
   // 타겟 정보 가져오기
-  let target: User | null = null;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === targetId) {
-      target = nowRoom.users[i];
+  let target: UserSessions | null = null;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === targetId) {
+      target = nowRoom.getUsers()[i];
       break;
     }
   }
@@ -893,23 +867,23 @@ const attackTarget = async (attackerId: number, roomId: number, skillCoeffcient:
   }
 
   // 공격 스킬 실행
-  const damage = Math.round(attacker.character.attack * skillCoeffcient - target.character.armor);
-  target.character.hp = Math.max(target.character.hp - damage, 0);
-  if (target.character.aliveState && target.character.hp <= 0) {
-    attacker.character.exp += target.character.exp;
-    await monsterReward(nowRooms, nowRoom, attacker, target);
+  const damage = Math.round(attacker.getCharacter().attack * skillCoeffcient - target.getCharacter().armor);
+  target.getCharacter().hp = Math.max(target.getCharacter().hp - damage, 0);
+  if (target.getCharacter().aliveState && target.getCharacter().hp <= 0) {
+    attacker.getCharacter().exp += target.getCharacter().exp;
+    monsterReward(nowRooms, nowRoom, attacker, target);
   }
 
   // 보스를 죽였는지 검사
-  if (target.character.roleType === RoleType.BOSS_MONSTER && target.character.hp <= 0) {
-    await gameEndNotification(roomId, 3);
+  if (target.getCharacter().roleType === RoleType.BOSS_MONSTER && target.getCharacter().hp <= 0) {
+    gameEndNotification(roomId, 3);
     return;
   }
 
   // 생존자를 죽였는지 검사
-  if (target.character.roleType === RoleType.SUR5VAL && target.character.hp <= 0) {
+  if (target.getCharacter().roleType === RoleType.SUR5VAL && target.getCharacter().hp <= 0) {
     for (let i = 0; i < shoppingUserIdSessions[roomId].length; i++) {
-      if (shoppingUserIdSessions[roomId][i][0] === attacker.id) {
+      if (shoppingUserIdSessions[roomId][i][0] === attacker.getId()) {
         shoppingUserIdSessions[roomId].splice(i, 1);
         break;
       }
@@ -917,31 +891,25 @@ const attackTarget = async (attackerId: number, roomId: number, skillCoeffcient:
     return;
   }
 
-  await setRedisData('roomData', nowRooms);
-  await userUpdateNotification(nowRoom);
+  userUpdateNotification(nowRoom);
 };
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 타겟이 된 적 공격 (방어력 무시 공격)
-const attackTargetTrueDamage = async (
-  attackerId: number,
-  roomId: number,
-  skillCoeffcient: number,
-  targetId: number
-) => {
+const attackTargetTrueDamage = (attackerId: number, roomId: number, skillCoeffcient: number, targetId: number) => {
   // 공격 실행 중 라운드가 바뀌지 않았는지 검사
-  const nowRooms = await getRedisData('roomData');
+  const nowRooms = Server.getInstance().getRooms();
   let isChanged: boolean = true;
-  let nowRoom: Room | null = null;
+  let nowRoom: GameRoom | null = null;
   for (let i = 0; i < nowRooms.length; i++) {
-    if (nowRooms[i].id === roomId) {
+    if (nowRooms[i].getRoomId() === roomId) {
       nowRoom = nowRooms[i];
       break;
     }
   }
   if (!nowRoom) return;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === targetId) isChanged = false;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === targetId) isChanged = false;
   }
   if (isChanged) {
     console.log('공격 실행 중 라운드 바뀌어서 몬스터 사라짐');
@@ -949,10 +917,10 @@ const attackTargetTrueDamage = async (
   }
 
   // 공격자의 정보 가져오기
-  let attacker: User | null = null;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === attackerId) {
-      attacker = nowRoom.users[i];
+  let attacker: UserSessions | null = null;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === attackerId) {
+      attacker = nowRoom.getUsers()[i];
       break;
     }
   }
@@ -962,10 +930,10 @@ const attackTargetTrueDamage = async (
   }
 
   // 타겟 정보 가져오기
-  let target: User | null = null;
-  for (let i = 0; i < nowRoom.users.length; i++) {
-    if (nowRoom.users[i].id === targetId) {
-      target = nowRoom.users[i];
+  let target: UserSessions | null = null;
+  for (let i = 0; i < nowRoom.getUsers().length; i++) {
+    if (nowRoom.getUsers()[i].getId() === targetId) {
+      target = nowRoom.getUsers()[i];
       break;
     }
   }
@@ -975,30 +943,28 @@ const attackTargetTrueDamage = async (
   }
 
   // 공격 스킬 실행
-  const damage = Math.round(attacker.character.attack * skillCoeffcient);
-  target.character.hp = Math.max(target.character.hp - damage, 0);
-  if (target.character.aliveState && target.character.hp <= 0) {
-    attacker.character.exp += target.character.exp;
-    await monsterReward(nowRooms, nowRoom, attacker, target);
+  const damage = Math.round(attacker.getCharacter().attack * skillCoeffcient);
+  target.getCharacter().hp = Math.max(target.getCharacter().hp - damage, 0);
+  if (target.getCharacter().aliveState && target.getCharacter().hp <= 0) {
+    attacker.getCharacter().exp += target.getCharacter().exp;
+    monsterReward(nowRooms, nowRoom, attacker, target);
   }
 
   // 보스를 죽였는지 검사
-  if (target.character.roleType === RoleType.BOSS_MONSTER && target.character.hp <= 0) {
-    await gameEndNotification(roomId, 3);
+  if (target.getCharacter().roleType === RoleType.BOSS_MONSTER && target.getCharacter().hp <= 0) {
+    gameEndNotification(roomId, 3);
     return;
   }
 
-  await setRedisData('roomData', nowRooms);
-  await userUpdateNotification(nowRoom);
+  userUpdateNotification(nowRoom);
 };
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 해당 캐릭터(아군) 버프
-const changeStatus = async (
+const changeStatus = (
   manaCost: number,
-  user: User,
-  rooms: Room[],
-  room: Room,
+  user: UserSessions,
+  room: GameRoom,
   hp: number,
   armor: number,
   attack: number,
@@ -1006,33 +972,32 @@ const changeStatus = async (
   cardType: number
 ) => {
   // 마나가 충분한지 검사
-  if (user.character.mp < manaCost) {
+  if (user.getCharacter().mp < manaCost) {
     console.log('마나가 부족합니다.');
     return;
   }
 
   // 이미 버프 상태인지 검사
-  if (characterBuffStatus[user.id]) {
-    for (let i = 0; i < characterBuffStatus[user.id].length; i++) {
-      if (characterBuffStatus[user.id][i] === cardType) {
+  if (characterBuffStatus[user.getId()]) {
+    for (let i = 0; i < characterBuffStatus[user.getId()].length; i++) {
+      if (characterBuffStatus[user.getId()][i] === cardType) {
         console.log('이미 버프 상태입니다.');
         return;
       }
     }
   } else {
-    characterBuffStatus[user.id] = [];
+    characterBuffStatus[user.getId()] = [];
   }
 
   // 버프 스킬 실행
-  characterBuffStatus[user.id].push(cardType);
-  user.character.mp -= manaCost;
-  user.character.hp += hp;
-  user.character.maxHp += hp;
-  user.character.attack += attack;
-  user.character.armor += armor;
+  characterBuffStatus[user.getId()].push(cardType);
+  user.getCharacter().mp -= manaCost;
+  user.getCharacter().hp += hp;
+  user.getCharacter().maxHp += hp;
+  user.getCharacter().attack += attack;
+  user.getCharacter().armor += armor;
   console.log('버프 사용 후: ', characterBuffStatus);
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
+  userUpdateNotification(room);
 
   // 버프 해제
   setTimeout(async () => {
@@ -1040,15 +1005,15 @@ const changeStatus = async (
     deleteBuff(user, cardType);
 
     // 버프 끝나는 시점의 room 정보 찾기
-    const nowRooms: Room[] | undefined = await getRedisData('roomData');
+    const nowRooms: GameRoom[] | undefined = Server.getInstance().getRooms();
     if (nowRooms === undefined) {
       console.error('서버에 Rooms정보가 존재하지 않습니다.');
       return;
     }
-    let nowRoom: Room | undefined;
+    let nowRoom: GameRoom | undefined;
     for (let i = 0; i < nowRooms.length; i++) {
-      for (let j = 0; j < nowRooms[i].users.length; j++) {
-        if (nowRooms[i].users[j].id === user.id) {
+      for (let j = 0; j < nowRooms[i].getUsers().length; j++) {
+        if (nowRooms[i].getUsers()[j].getId() === user.getId()) {
           nowRoom = nowRooms[i];
         }
       }
@@ -1059,13 +1024,13 @@ const changeStatus = async (
     }
 
     // 버프 해제 시점에 기존 게임이 종료된 상태일 경우
-    if (nowRoom.id !== room.id) return;
+    if (nowRoom.getRoomId() !== room.getRoomId()) return;
 
     // 버프 해제 시점의 유저 찾기
-    let skillUser: User | null = null;
-    for (let i = 0; i < nowRoom.users.length; i++) {
-      if (nowRoom.users[i].id === user.id) {
-        skillUser = nowRoom.users[i];
+    let skillUser: UserSessions | null = null;
+    for (let i = 0; i < nowRoom.getUsers().length; i++) {
+      if (nowRoom.getUsers()[i].getId() === user.getId()) {
+        skillUser = nowRoom.getUsers()[i];
         break;
       }
     }
@@ -1075,14 +1040,13 @@ const changeStatus = async (
     }
 
     // 버프 해제 실행
-    skillUser.character.hp -= hp;
-    skillUser.character.maxHp -= hp;
-    skillUser.character.attack -= attack;
-    skillUser.character.armor -= armor;
-    skillUser.character.hp = Math.min(skillUser.character.hp, skillUser.character.maxHp);
+    skillUser.getCharacter().hp -= hp;
+    skillUser.getCharacter().maxHp -= hp;
+    skillUser.getCharacter().attack -= attack;
+    skillUser.getCharacter().armor -= armor;
+    skillUser.getCharacter().hp = Math.min(skillUser.getCharacter().hp, skillUser.getCharacter().maxHp);
 
-    await setRedisData('roomData', nowRooms);
-    await userUpdateNotification(nowRoom);
+    userUpdateNotification(nowRoom);
   }, buffTime);
 
   return true;
@@ -1090,11 +1054,10 @@ const changeStatus = async (
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 파티 버프 (모든 아군 캐릭터 능력치 변동)
-const partyBuff = async (
+const partyBuff = (
   manaCost: number,
-  user: User,
-  rooms: Room[],
-  room: Room,
+  user: UserSessions,
+  room: GameRoom,
   hp: number,
   armor: number,
   attack: number,
@@ -1102,37 +1065,36 @@ const partyBuff = async (
   cardType: number
 ) => {
   // 마나가 충분한지 검사
-  if (user.character.mp < manaCost) {
+  if (user.getCharacter().mp < manaCost) {
     console.log('마나가 부족합니다.');
     return;
   }
 
   // 이미 버프 상태인지 검사
-  if (characterBuffStatus[user.id]) {
-    for (let i = 0; i < characterBuffStatus[user.id].length; i++) {
-      if (characterBuffStatus[user.id][i] === cardType) {
+  if (characterBuffStatus[user.getId()]) {
+    for (let i = 0; i < characterBuffStatus[user.getId()].length; i++) {
+      if (characterBuffStatus[user.getId()][i] === cardType) {
         console.log('이미 버프 상태입니다.');
         return;
       }
     }
   } else {
-    characterBuffStatus[user.id] = [];
+    characterBuffStatus[user.getId()] = [];
   }
 
   // 버프 스킬 실행
-  characterBuffStatus[user.id].push(cardType);
-  user.character.mp -= manaCost;
-  for (let i = 0; i < room.users.length; i++) {
-    if (room.users[i].character.roleType === RoleType.SUR5VAL) {
-      room.users[i].character.hp += hp;
-      room.users[i].character.maxHp += hp;
-      room.users[i].character.attack += attack;
-      room.users[i].character.armor += armor;
+  characterBuffStatus[user.getId()].push(cardType);
+  user.getCharacter().mp -= manaCost;
+  for (let i = 0; i < room.getUsers().length; i++) {
+    if (room.getUsers()[i].getCharacter().roleType === RoleType.SUR5VAL) {
+      room.getUsers()[i].getCharacter().hp += hp;
+      room.getUsers()[i].getCharacter().maxHp += hp;
+      room.getUsers()[i].getCharacter().attack += attack;
+      room.getUsers()[i].getCharacter().armor += armor;
     }
   }
   console.log('버프 사용 후: ', characterBuffStatus);
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
+  userUpdateNotification(room);
 
   // 버프 해제
   setTimeout(async () => {
@@ -1140,15 +1102,15 @@ const partyBuff = async (
     deleteBuff(user, cardType);
 
     // 버프 끝나는 시점의 room 정보 찾기
-    const nowRooms: Room[] | undefined = await getRedisData('roomData');
+    const nowRooms: GameRoom[] | undefined = Server.getInstance().getRooms();
     if (nowRooms === undefined) {
       console.error('서버에 Rooms정보가 존재하지 않습니다.');
       return;
     }
-    let nowRoom: Room | undefined;
+    let nowRoom: GameRoom | undefined;
     for (let i = 0; i < nowRooms.length; i++) {
-      for (let j = 0; j < nowRooms[i].users.length; j++) {
-        if (nowRooms[i].users[j].id === user.id) {
+      for (let j = 0; j < nowRooms[i].getUsers().length; j++) {
+        if (nowRooms[i].getUsers()[j].getId() === user.getId()) {
           nowRoom = nowRooms[i];
         }
       }
@@ -1159,24 +1121,26 @@ const partyBuff = async (
     }
 
     // 버프 해제 시점에 기존 게임이 종료된 상태일 경우
-    if (nowRoom.id !== room.id) {
+    if (nowRoom.getRoomId() !== room.getRoomId()) {
       console.error('게임이 이미 종료되었습니다.');
       return;
     }
 
     // 버프 해제 실행
-    for (let i = 0; i < nowRoom.users.length; i++) {
-      if (nowRoom.users[i].character.roleType === RoleType.SUR5VAL) {
-        nowRoom.users[i].character.hp -= hp;
-        nowRoom.users[i].character.maxHp -= hp;
-        nowRoom.users[i].character.attack -= attack;
-        nowRoom.users[i].character.armor -= armor;
-        nowRoom.users[i].character.hp = Math.min(nowRoom.users[i].character.hp, nowRoom.users[i].character.maxHp);
+    for (let i = 0; i < nowRoom.getUsers().length; i++) {
+      if (nowRoom.getUsers()[i].getCharacter().roleType === RoleType.SUR5VAL) {
+        nowRoom.getUsers()[i].getCharacter().hp -= hp;
+        nowRoom.getUsers()[i].getCharacter().maxHp -= hp;
+        nowRoom.getUsers()[i].getCharacter().attack -= attack;
+        nowRoom.getUsers()[i].getCharacter().armor -= armor;
+        nowRoom.getUsers()[i].getCharacter().hp = Math.min(
+          nowRoom.getUsers()[i].getCharacter().hp,
+          nowRoom.getUsers()[i].getCharacter().maxHp
+        );
       }
     }
 
-    await setRedisData('roomData', nowRooms);
-    await userUpdateNotification(nowRoom);
+    userUpdateNotification(nowRoom);
   }, buffTime);
 };
 
@@ -1184,9 +1148,8 @@ const partyBuff = async (
 // 자신 주변을 일정시간동안 공격하는 버프 (정령 소환 버프)
 const summonSpiritBuff = async (
   manaCost: number,
-  attacker: User,
-  room: Room,
-  rooms: Room[],
+  attacker: UserSessions,
+  room: GameRoom,
   skillCoeffcient: number,
   range: number,
   targetNumber: number,
@@ -1199,35 +1162,34 @@ const summonSpiritBuff = async (
   if (!initGameInfo) return;
 
   const characterAttackCool = initGameInfo.coolTime;
-  if (Date.now() - attacker.character.coolDown < characterAttackCool) {
+  if (Date.now() - attacker.getCharacter().coolDown < characterAttackCool) {
     console.log('공격 쿨타임 중입니다.');
     return;
   }
 
   // 마나가 충분한지 검사
-  if (attacker.character.mp < manaCost) {
+  if (attacker.getCharacter().mp < manaCost) {
     console.log('마나가 부족합니다.');
     return;
   }
 
   // 이미 버프 상태인지 검사
-  if (characterBuffStatus[attacker.id]) {
-    for (let i = 0; i < characterBuffStatus[attacker.id].length; i++) {
-      if (characterBuffStatus[attacker.id][i] === cardType) {
+  if (characterBuffStatus[attacker.getId()]) {
+    for (let i = 0; i < characterBuffStatus[attacker.getId()].length; i++) {
+      if (characterBuffStatus[attacker.getId()][i] === cardType) {
         console.log('이미 버프 상태입니다.');
         return;
       }
     }
   } else {
-    characterBuffStatus[attacker.id] = [];
+    characterBuffStatus[attacker.getId()] = [];
   }
 
   // 버프 상태 추가
-  characterBuffStatus[attacker.id].push(cardType);
+  characterBuffStatus[attacker.getId()].push(cardType);
 
   // 마나 소모
-  attacker.character.mp -= manaCost;
-  await setRedisData('roomData', rooms);
+  attacker.getCharacter().mp -= manaCost;
   userUpdateNotification(room);
 
   // 스킬 정보 생성
@@ -1247,17 +1209,17 @@ const summonSpiritBuff = async (
     }
 
     // 스킬이 끝나기 전에 게임이 종료되었는지 확인 및 종료
-    const nowRooms: Room[] | undefined = await getRedisData('roomData');
+    const nowRooms: GameRoom[] | undefined = Server.getInstance().getRooms();
     if (nowRooms === undefined) {
       deleteBuff(attacker, cardType);
       clearInterval(attackSkill);
       console.log('스킬 사용자가 속한 room 정보를 찾을 수 없습니다.');
       return;
     }
-    let nowRoom: Room | undefined;
+    let nowRoom: GameRoom | undefined;
     for (let i = 0; i < nowRooms.length; i++) {
-      for (let j = 0; j < nowRooms[i].users.length; j++) {
-        if (nowRooms[i].users[j].id === attacker.id) {
+      for (let j = 0; j < nowRooms[i].getUsers().length; j++) {
+        if (nowRooms[i].getUsers()[j].getId() === attacker.getId()) {
           nowRoom = nowRooms[i];
         }
       }
@@ -1268,7 +1230,7 @@ const summonSpiritBuff = async (
       console.log('스킬 사용자가 속한 room 정보를 찾을 수 없습니다.');
       return;
     }
-    if (nowRoom.id !== room.id) {
+    if (nowRoom.getRoomId() !== room.getRoomId()) {
       deleteBuff(attacker, cardType);
       clearInterval(attackSkill);
       console.log('게임이 이미 종료되어서 스킬을 종료합니다.');
@@ -1276,27 +1238,33 @@ const summonSpiritBuff = async (
     }
 
     // 공격자의 상태 값 얻기
-    let skillUser: User | null = null;
-    for (let i = 0; i < nowRoom.users.length; i++) {
-      if (nowRoom.users[i].id === attacker.id) {
-        skillUser = nowRoom.users[i];
+    let skillUser: UserSessions | null = null;
+    for (let i = 0; i < nowRoom.getUsers().length; i++) {
+      if (nowRoom.getUsers()[i].getId() === attacker.getId()) {
+        skillUser = nowRoom.getUsers()[i];
       }
     }
     if (!skillUser) return;
 
     // 공격자가 죽어있으면 종료
-    if (skillUser.character.hp <= 0) {
+    if (skillUser.getCharacter().hp <= 0) {
       deleteBuff(attacker, cardType);
       clearInterval(attackSkill);
       console.log('버프 스킬 시전자가 사망하여 정령 버프가 종료되었습니다.');
     }
 
     // 공격자의 위치 값 얻기
-    const characterPositionDatas = await getRedisData('characterPositionDatas');
-    const characterPositionData: CharacterPositionData[] = characterPositionDatas[room.id];
+    const positionSessions = Server.getInstance()
+      .getPositions()
+      .find((PositionSessions) => PositionSessions.getPositionRoomId() === room.getRoomId());
+    if (!positionSessions) {
+      console.error('characterPositionData를 찾지 못하였습니다.');
+      return;
+    }
+    const characterPositionData = positionSessions.getCharacterPositions();
     let attackerPosition: CharacterPositionData | null = null;
     for (let i = 0; i < characterPositionData.length; i++) {
-      if (characterPositionData[i].id === attacker.id) {
+      if (characterPositionData[i].id === attacker.getId()) {
         attackerPosition = characterPositionData[i];
         break;
       }
@@ -1308,15 +1276,15 @@ const summonSpiritBuff = async (
 
     const targetMonsters = [];
     // 체력 0 초과의 적 정보 얻기
-    for (let i = 0; i < nowRoom.users.length; i++) {
+    for (let i = 0; i < nowRoom.getUsers().length; i++) {
       if (
-        (nowRoom.users[i].character.roleType === RoleType.WEAK_MONSTER ||
-          nowRoom.users[i].character.roleType === RoleType.BOSS_MONSTER) &&
-        nowRoom.users[i].character.hp > 0
+        (nowRoom.getUsers()[i].getCharacter().roleType === RoleType.WEAK_MONSTER ||
+          nowRoom.getUsers()[i].getCharacter().roleType === RoleType.BOSS_MONSTER) &&
+        nowRoom.getUsers()[i].getCharacter().hp > 0
       ) {
         let enemyPosition: CharacterPositionData | null = null;
         for (let j = 0; j < characterPositionData.length; j++) {
-          if (characterPositionData[j].id === nowRoom.users[i].id) {
+          if (characterPositionData[j].id === nowRoom.getUsers()[i].getId()) {
             enemyPosition = characterPositionData[j];
             break;
           }
@@ -1329,7 +1297,7 @@ const summonSpiritBuff = async (
         }
         if ((enemyPosition.x - attackerPosition.x) ** 2 + (enemyPosition.y - attackerPosition.y) ** 2 < range ** 2) {
           targetMonsters.push({
-            monster: nowRoom.users[i],
+            monster: nowRoom.getUsers()[i],
             distance: (enemyPosition.x - attackerPosition.x) ** 2 + (enemyPosition.y - attackerPosition.y) ** 2
           });
         }
@@ -1346,30 +1314,29 @@ const summonSpiritBuff = async (
     // 공격 실행
     for (let i = 0; i < targetMonsters.length; i++) {
       // 딜링 및 애니메이션 재생
-      const target = targetMonsters[i].monster.character;
-      target.hp = Math.max(target.hp - skillUser.character.attack * skillCoeffcient + target.armor, 0);
+      const target = targetMonsters[i].monster.getCharacter();
+      target.hp = Math.max(target.hp - skillUser.getCharacter().attack * skillCoeffcient + target.armor, 0);
       sendAnimation(room, targetMonsters[i].monster, 7);
-      for (let j = 0; j < monsterAiDatas[room.id].length; j++) {
-        if (monsterAiDatas[room.id][j].id === targetMonsters[i].monster.id) {
-          monsterAiDatas[room.id][j].animationDelay = 5;
+      for (let j = 0; j < monsterAiDatas[room.getRoomId()].length; j++) {
+        if (monsterAiDatas[room.getRoomId()][j].id === targetMonsters[i].monster.getId()) {
+          monsterAiDatas[room.getRoomId()][j].animationDelay = 5;
           break;
         }
       }
 
       // 처치 검사 하여 보상 획득
       if (target.hp <= 0 && target.aliveState) {
-        skillUser.character.exp += target.exp;
-        await monsterReward(nowRooms, nowRoom, skillUser, targetMonsters[i].monster);
+        skillUser.getCharacter().exp += target.exp;
+        monsterReward(nowRooms, nowRoom, skillUser, targetMonsters[i].monster);
       }
 
       // 보스를 죽였는지 검사
       if (target.roleType === RoleType.BOSS_MONSTER && target.hp <= 0) {
-        await gameEndNotification(nowRoom.id, 3);
+        gameEndNotification(nowRoom.getRoomId(), 3);
         return;
       }
     }
     lastAttack = Date.now();
-    await setRedisData('roomData', nowRooms);
     await userUpdateNotification(nowRoom);
   }, 200);
 
@@ -1378,12 +1345,12 @@ const summonSpiritBuff = async (
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 본인의 Hp, Mp 회복 / Exp 획득
-const usePotion = async (user: User, rooms: Room[], room: Room, cardsType: number) => {
+const usePotion = (user: UserSessions, room: GameRoom, cardsType: number) => {
   // 해당 아이템 보유여부 검사 및 개수 차감
   let isOwned: boolean = false;
-  for (let i = 0; i < user.character.handCards.length; i++) {
-    if (user.character.handCards[i].type === cardsType && user.character.handCards[i].count > 0) {
-      user.character.handCards[i].count--;
+  for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+    if (user.getCharacter().handCards[i].type === cardsType && user.getCharacter().handCards[i].count > 0) {
+      user.getCharacter().handCards[i].count--;
       isOwned = true;
       break;
     }
@@ -1410,20 +1377,20 @@ const usePotion = async (user: User, rooms: Room[], room: Room, cardsType: numbe
     console.error('유저 캐릭터 초기 정보를 찾을 수 없습니다.');
     return;
   }
-  const characterStatInfo = characterStatInfos.find((data) => data.characterType === user.character.characterType);
+  const characterStatInfo = characterStatInfos.find((data) => data.characterType === user.getCharacter().characterType);
   if (!characterStatInfo) {
     console.error('유저 캐릭터 초기 정보를 찾을 수 없습니다.');
     return;
   }
-  user.character.hp = Math.min(user.character.hp + potionData.hp, user.character.maxHp);
-  user.character.mp = Math.min(user.character.mp + potionData.mp, characterStatInfo.mp);
-  user.character.exp += potionData.exp;
+  user.getCharacter().hp = Math.min(user.getCharacter().hp + potionData.hp, user.getCharacter().maxHp);
+  user.getCharacter().mp = Math.min(user.getCharacter().mp + potionData.mp, characterStatInfo.mp);
+  user.getCharacter().exp += potionData.exp;
 
   // 레벨업 확인
-  while (user.character.exp >= user.character.maxExp) {
-    user.character.exp -= user.character.maxExp;
-    user.character.level += 1;
-    user.character.maxExp += 10;
+  while (user.getCharacter().exp >= user.getCharacter().maxExp) {
+    user.getCharacter().exp -= user.getCharacter().maxExp;
+    user.getCharacter().level += 1;
+    user.getCharacter().maxExp += 10;
     // 직업별 스탯 증가
     if (!setStatRewards(user)) {
       console.error('setStatRewards 실패');
@@ -1436,13 +1403,18 @@ const usePotion = async (user: User, rooms: Room[], room: Room, cardsType: numbe
     }
   }
 
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
+  userUpdateNotification(room);
 };
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 방어구 장착 equipType (0: 투구, 1: 갑옷, 2: 망토, 3: 장갑)
-const equipItem = async (user: User, rooms: Room[], room: Room, equipIndex: number, cardsType: number) => {
+const equipItem = (
+  user: UserSessions,
+
+  room: GameRoom,
+  equipIndex: number,
+  cardsType: number
+) => {
   // DB의 카드 생성 정보 가져오기
   const equipCardDBInfo = Server.getInstance().equipItemInfo;
   if (!equipCardDBInfo) {
@@ -1457,9 +1429,9 @@ const equipItem = async (user: User, rooms: Room[], room: Room, equipIndex: numb
 
   // 해당 아이템 보유여부 검사
   let isOwned: boolean = false;
-  for (let i = 0; i < user.character.handCards.length; i++) {
-    if (user.character.handCards[i].type === cardsType && user.character.handCards[i].count > 0) {
-      user.character.handCards[i].count--;
+  for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+    if (user.getCharacter().handCards[i].type === cardsType && user.getCharacter().handCards[i].count > 0) {
+      user.getCharacter().handCards[i].count--;
       isOwned = true;
       break;
     }
@@ -1472,38 +1444,41 @@ const equipItem = async (user: User, rooms: Room[], room: Room, equipIndex: numb
   }
 
   // 장착중이던 방어구 해제
-  if (![302, 303, 304, 305].includes(user.character.equips[equipIndex]) && user.character.equips[equipIndex]) {
+  if (
+    ![302, 303, 304, 305].includes(user.getCharacter().equips[equipIndex]) &&
+    user.getCharacter().equips[equipIndex]
+  ) {
     // 장착칸 → handCards로 옮기기
     let isInHandCards: boolean = false;
-    for (let i = 0; i < user.character.handCards.length; i++) {
-      if (user.character.handCards[i].type === user.character.equips[equipIndex]) {
-        user.character.handCards[i].count++;
+    for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+      if (user.getCharacter().handCards[i].type === user.getCharacter().equips[equipIndex]) {
+        user.getCharacter().handCards[i].count++;
         isInHandCards = true;
         break;
       }
     }
-    if (isInHandCards === false) user.character.handCards.push({ type: user.character.equips[equipIndex], count: 1 });
+    if (isInHandCards === false)
+      user.getCharacter().handCards.push({ type: user.getCharacter().equips[equipIndex], count: 1 });
 
     // 능력치 하락
-    const prevEquipCard = equipCardDBInfo.find((data) => data.cardType === user.character.equips[equipIndex]);
+    const prevEquipCard = equipCardDBInfo.find((data) => data.cardType === user.getCharacter().equips[equipIndex]);
     if (prevEquipCard) {
-      user.character.attack -= prevEquipCard.attack;
-      user.character.armor -= prevEquipCard.armor;
-      user.character.maxHp -= prevEquipCard.hp;
-      user.character.hp -= prevEquipCard.hp;
+      user.getCharacter().attack -= prevEquipCard.attack;
+      user.getCharacter().armor -= prevEquipCard.armor;
+      user.getCharacter().maxHp -= prevEquipCard.hp;
+      user.getCharacter().hp -= prevEquipCard.hp;
     }
   }
 
   // 새로운 방어구 장착 및 능력치 상승
-  user.character.equips[equipIndex] = cardsType;
-  user.character.attack += equipCard.attack;
-  user.character.armor += equipCard.armor;
-  user.character.maxHp += equipCard.hp;
-  user.character.hp += equipCard.hp;
+  user.getCharacter().equips[equipIndex] = cardsType;
+  user.getCharacter().attack += equipCard.attack;
+  user.getCharacter().armor += equipCard.armor;
+  user.getCharacter().maxHp += equipCard.hp;
+  user.getCharacter().hp += equipCard.hp;
 
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
-  sendPacket(socketSessions[user.id], config.packetType.USE_CARD_RESPONSE, {
+  userUpdateNotification(room);
+  sendPacket(socketSessions[user.getId()], config.packetType.USE_CARD_RESPONSE, {
     success: true,
     failCode: 0
   });
@@ -1511,7 +1486,7 @@ const equipItem = async (user: User, rooms: Room[], room: Room, equipIndex: numb
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 무기 장착
-const equipWeapon = async (user: User, rooms: Room[], room: Room, cardsType: number) => {
+const equipWeapon = (user: UserSessions, room: GameRoom, cardsType: number) => {
   // DB의 카드 생성 정보 가져오기
   const equipCardDBInfo = Server.getInstance().equipItemInfo;
   if (!equipCardDBInfo) {
@@ -1525,9 +1500,9 @@ const equipWeapon = async (user: User, rooms: Room[], room: Room, cardsType: num
   }
   // 해당 아이템 보유여부 검사
   let isOwned: boolean = false;
-  for (let i = 0; i < user.character.handCards.length; i++) {
-    if (user.character.handCards[i].type === cardsType && user.character.handCards[i].count > 0) {
-      user.character.handCards[i].count--;
+  for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+    if (user.getCharacter().handCards[i].type === cardsType && user.getCharacter().handCards[i].count > 0) {
+      user.getCharacter().handCards[i].count--;
       isOwned = true;
       break;
     }
@@ -1535,38 +1510,37 @@ const equipWeapon = async (user: User, rooms: Room[], room: Room, cardsType: num
   if (isOwned === false) return;
 
   // // 장착중이던 무기 해제
-  if (user.character.weapon !== 301 && user.character.weapon) {
+  if (user.getCharacter().weapon !== 301 && user.getCharacter().weapon) {
     // 장착칸 → handCards로 옮기기
     let isInHandCards: boolean = false;
-    for (let i = 0; i < user.character.handCards.length; i++) {
-      if (user.character.handCards[i].type === user.character.weapon) {
-        user.character.handCards[i].count++;
+    for (let i = 0; i < user.getCharacter().handCards.length; i++) {
+      if (user.getCharacter().handCards[i].type === user.getCharacter().weapon) {
+        user.getCharacter().handCards[i].count++;
         isInHandCards = true;
         break;
       }
     }
-    if (isInHandCards === false) user.character.handCards.push({ type: user.character.weapon, count: 1 });
+    if (isInHandCards === false) user.getCharacter().handCards.push({ type: user.getCharacter().weapon, count: 1 });
 
     // 능력치 하락
-    const prevEquipCard = equipCardDBInfo.find((data) => data.cardType === user.character.weapon);
+    const prevEquipCard = equipCardDBInfo.find((data) => data.cardType === user.getCharacter().weapon);
     if (prevEquipCard) {
-      user.character.attack -= prevEquipCard.attack;
-      user.character.armor -= prevEquipCard.armor;
-      user.character.maxHp -= prevEquipCard.hp;
-      user.character.hp -= prevEquipCard.hp;
+      user.getCharacter().attack -= prevEquipCard.attack;
+      user.getCharacter().armor -= prevEquipCard.armor;
+      user.getCharacter().maxHp -= prevEquipCard.hp;
+      user.getCharacter().hp -= prevEquipCard.hp;
     }
   }
 
   // 새로운 무기 장착 및 능력치 상승
-  user.character.weapon = cardsType;
-  user.character.attack += weaponCard.attack;
-  user.character.armor += weaponCard.armor;
-  user.character.maxHp += weaponCard.hp;
-  user.character.hp += weaponCard.hp;
+  user.getCharacter().weapon = cardsType;
+  user.getCharacter().attack += weaponCard.attack;
+  user.getCharacter().armor += weaponCard.armor;
+  user.getCharacter().maxHp += weaponCard.hp;
+  user.getCharacter().hp += weaponCard.hp;
 
-  await setRedisData('roomData', rooms);
-  await userUpdateNotification(room);
-  sendPacket(socketSessions[user.id], config.packetType.USE_CARD_RESPONSE, {
+  userUpdateNotification(room);
+  sendPacket(socketSessions[user.getId()], config.packetType.USE_CARD_RESPONSE, {
     success: true,
     failCode: 0
   });
@@ -1574,13 +1548,13 @@ const equipWeapon = async (user: User, rooms: Room[], room: Room, cardsType: num
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 버프 사용자 목록에서 제거
-const deleteBuff = (user: User, buffType: number) => {
-  for (let i = 0; i < characterBuffStatus[user.id].length; i++) {
-    if (characterBuffStatus[user.id][i] === buffType) {
-      characterBuffStatus[user.id].splice(i, 1);
-      console.log('삭제 후 남은 버프', characterBuffStatus[user.id]);
-      if (characterBuffStatus[user.id].length === 0) {
-        delete characterBuffStatus[user.id];
+const deleteBuff = (user: UserSessions, buffType: number) => {
+  for (let i = 0; i < characterBuffStatus[user.getId()].length; i++) {
+    if (characterBuffStatus[user.getId()][i] === buffType) {
+      characterBuffStatus[user.getId()].splice(i, 1);
+      console.log('삭제 후 남은 버프', characterBuffStatus[user.getId()]);
+      if (characterBuffStatus[user.getId()].length === 0) {
+        delete characterBuffStatus[user.getId()];
         break;
       }
     }
@@ -1589,25 +1563,25 @@ const deleteBuff = (user: User, buffType: number) => {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 공격 유효성 검증
-const attackPossible = async (cardType: CardType, attacker: User, target: User, needMp: number) => {
+const attackPossible = async (cardType: CardType, attacker: UserSessions, target: UserSessions, needMp: number) => {
   // 캐릭터의 공격 쿨타임 검사
   const initGameInfo: skillCardDBData = await dbManager.skillCardInfo(cardType);
   if (!initGameInfo) return false;
 
   const characterAttackCool = initGameInfo.coolTime;
-  if (Date.now() - attacker.character.coolDown < characterAttackCool) {
+  if (Date.now() - attacker.getCharacter().coolDown < characterAttackCool) {
     console.log('공격 쿨타임 중입니다.');
     return false;
   }
-  if (attacker.character.mp < needMp) {
+  if (attacker.getCharacter().mp < needMp) {
     console.log('마나가 부족합니다.');
     return false;
   }
-  if (target.character.roleType === RoleType.SUR5VAL) {
+  if (target.getCharacter().roleType === RoleType.SUR5VAL) {
     console.log('아군을 공격할 수는 없습니다.');
     return false;
   }
-  if (target.character.hp <= 0) {
+  if (target.getCharacter().hp <= 0) {
     console.log('살아있는 적만 공격할 수 있습니다.');
     return false;
   }
@@ -1617,12 +1591,12 @@ const attackPossible = async (cardType: CardType, attacker: User, target: User, 
 
 /////////////////////////////////////////////////////////////////////////////////////////////////
 // 스킬 애니메이션 패킷 보내기
-const sendAnimation = (room: Room, animationTarget: User, animationType: number) => {
-  for (let i = 0; i < room.users.length; i++) {
-    const socket: CustomSocket | undefined = socketSessions[room.users[i].id];
+const sendAnimation = (room: GameRoom, animationTarget: UserSessions, animationType: number) => {
+  for (let i = 0; i < room.getUsers().length; i++) {
+    const socket: CustomSocket | undefined = socketSessions[room.getUsers()[i].getId()];
     if (socket) {
       sendPacket(socket, config.packetType.ANIMATION_NOTIFICATION, {
-        userId: animationTarget.id,
+        userId: animationTarget.getId(),
         animationType: animationType
       });
     }
