+++ src/gameServer/handlers/market/fleaMarketItemBuyHandler.ts
@@ -1,70 +1,76 @@
-import { Card, CustomSocket, FleaMarketItemSelectPayload, Room, User } from '../../../gameServer/interface/interface.js';
-import { getRedisData, getUserIdBySocket, setRedisData } from '../handlerMethod.js';
+import { Card, CustomSocket, FleaMarketItemSelectPayload } from '../../../gameServer/interface/interface.js';
+import { getUserBySocket } from '../handlerMethod.js';
 import { sendPacket } from '../../../packet/createPacket.js';
 import { config } from '../../../config/config.js';
 import { shoppingUserIdSessions } from '../../session/shoppingSession.js';
-import { userUpdateNotification } from '../notification/userUpdate.js';
 import Server from '../../class/server.js';
+import UserSessions from '../../class/userSessions.js';
+import GameRoom from '../../class/room.js';
+import { userUpdateNotification } from '../notification/userUpdate.js';
 
-export const fleaMarketItemBuy = async (socket: CustomSocket, payload: Object) => {
+export const fleaMarketItemBuy = (socket: CustomSocket, payload: Object) => {
+  const fleaMarketItemSelectPayload = payload as FleaMarketItemSelectPayload;
+
+  // 장비 데이터 가져오기
   const equipCardDBInfo = Server.getInstance().equipItemInfo;
   if (!equipCardDBInfo) {
     console.error('장비아이템 데이터가 없습니다.');
     return;
   }
+
+  // 소비 데이터 가져오기
   const consumableItemInfo = Server.getInstance().consumableItemInfo;
   if (!consumableItemInfo) {
     console.error('소비아이템 데이터가 없습니다.');
     return;
   }
 
-  const fleaMarketItemSelectPayload = payload as FleaMarketItemSelectPayload;
-
-  const userId: number | null = await getUserIdBySocket(socket);
-  if (!userId) {
-    console.error('fleaMarketItemSelect 레디스에 유저가 없음');
+  // 유저 데이터 가져오기 (userId 용도)
+  const user: UserSessions | null | undefined = getUserBySocket(socket);
+  if (!user) {
+    console.error('구매 시도하는 유저의 유저 정보가 없음');
     return;
   }
 
-  const rooms: Room[] = await getRedisData('roomData');
-  const room = rooms.find((room) => room.users.some((user) => user.id === userId));
-
+  // 유저의 roomData 가져오기
+  const rooms: GameRoom[] = Server.getInstance().getRooms();
+  const room = rooms.find((room) => room.getUsers().some((user) => user.getId() === user.getId()));
   if (!room) {
     console.error('fleaMarketItemSelect 방이 없음');
     return;
   }
 
-  let cardPickUser: User | null = null;
-  for (let i = 0; i < room.users.length; i++) {
-    if (room.users[i].id === userId) {
-      cardPickUser = room.users[i];
+  // 유저의 데이터 가져오기
+  let cardPickUser: UserSessions | null = null;
+  for (let i = 0; i < room.getUsers().length; i++) {
+    if (room.getUsers()[i].getId() === user.getId()) {
+      cardPickUser = room.getUsers()[i];
       break;
     }
   }
-
   if (cardPickUser === null) {
     console.error('fleaMarketItemSelect 유저가 없음');
     return;
   }
 
-  if (!cardPickUser.character || cardPickUser.character.hp === 0 || !cardPickUser.character.handCards) {
+  // 캐릭터 유효성 검사
+  if (!cardPickUser.getCharacter() || cardPickUser.getCharacter().hp === 0 || !cardPickUser.getCharacter().handCards) {
     console.error('fleaMarketItemSelect 캐릭터 에러');
     return;
   }
 
-  let redisFleaMarketCards: { [roomId: number]: number[] } | undefined = await getRedisData('fleaMarketCards');
-  if (!redisFleaMarketCards) {
-    console.error('fleaMarketItemSelect 레디스에 상점 카드가 없음');
+  // market 가져오기
+  const market = Server.getInstance()
+    .getMarkets()
+    .find((market) => market.getRoomId() === room.getRoomId());
+  if (!market) {
+    console.error('마켓 정보가저오기 실패');
     return;
   }
+  const marketItemsLists = market.getItemLists();
 
-  const cards = redisFleaMarketCards[room.id];
-  if (!cards || cards.length === 0 || cards.length < fleaMarketItemSelectPayload.pickIndex) {
-    console.error('fleaMarketItemSelect 레디스 상점 카드에 선택한 카드가 없음');
-    return;
-  }
-
-  const fleMarketPickCard = cards[fleaMarketItemSelectPayload.pickIndex];
+  // 선택된 카드 가져오기
+  const fleMarketPickCard = marketItemsLists[fleaMarketItemSelectPayload.pickIndex];
   if (!fleMarketPickCard) {
     console.error('fleaMarketItemSelect PickCard empty');
     return;
@@ -73,12 +79,12 @@ export const fleaMarketItemBuy = async (socket: CustomSocket, payload: Object) =
   // Exit 버튼 눌렀을 경우
   if (fleMarketPickCard === 1000) {
     sendPacket(socket, config.packetType.FLEA_MARKET_CARD_PICK_RESPONSE, {
-      userId: cardPickUser.id,
-      handCards: cardPickUser.character.handCards
+      userId: cardPickUser.getId(),
+      handCards: cardPickUser.getCharacter().handCards
     });
-    for (let i = 0; i < shoppingUserIdSessions[room.id].length; i++) {
-      if (shoppingUserIdSessions[room.id][i][0] === userId) {
-        shoppingUserIdSessions[room.id].splice(i, 1);
+    for (let i = 0; i < shoppingUserIdSessions[room.getRoomId()].length; i++) {
+      if (shoppingUserIdSessions[room.getRoomId()][i][0] === user.getId()) {
+        shoppingUserIdSessions[room.getRoomId()].splice(i, 1);
         break;
       }
     }
@@ -93,14 +99,14 @@ export const fleaMarketItemBuy = async (socket: CustomSocket, payload: Object) =
     console.error('카드 가격을 알 수 없습니다.(해당 카드가 데이터에 존재하지 않습니다.');
     return;
   }
-  if (cardPickUser.character.gold < pickedCardPrice) {
+  if (cardPickUser.getCharacter().gold < pickedCardPrice) {
     sendPacket(socket, config.packetType.FLEA_MARKET_CARD_PICK_RESPONSE, {
-      userId: cardPickUser.id,
-      handCards: cardPickUser.character.handCards
+      userId: cardPickUser.getId(),
+      handCards: cardPickUser.getCharacter().handCards
     });
-    for (let i = 0; i < shoppingUserIdSessions[room.id].length; i++) {
-      if (shoppingUserIdSessions[room.id][i][0] === userId) {
-        shoppingUserIdSessions[room.id].splice(i, 1);
+    for (let i = 0; i < shoppingUserIdSessions[room.getRoomId()].length; i++) {
+      if (shoppingUserIdSessions[room.getRoomId()][i][0] === user.getId()) {
+        shoppingUserIdSessions[room.getRoomId()].splice(i, 1);
         break;
       }
     }
@@ -114,30 +120,28 @@ export const fleaMarketItemBuy = async (socket: CustomSocket, payload: Object) =
     count: 1
   };
   let isOwned: boolean = false;
-  for (let i = 0; i < cardPickUser.character.handCards.length; i++) {
-    if (cardPickUser.character.handCards[i].type === newCard.type) {
-      cardPickUser.character.handCards[i].count++;
+  for (let i = 0; i < cardPickUser.getCharacter().handCards.length; i++) {
+    if (cardPickUser.getCharacter().handCards[i].type === newCard.type) {
+      cardPickUser.getCharacter().handCards[i].count++;
       isOwned = true;
       break;
     }
   }
-  if (isOwned === false) cardPickUser.character.handCards.push(newCard);
+  if (isOwned === false) cardPickUser.getCharacter().handCards.push(newCard);
 
   // cardType 순서대로 인벤토리 정렬
-  cardPickUser.character.handCards.sort((a, b) => a.type - b.type);
-  cardPickUser.character.gold -= pickedCardPrice;
+  cardPickUser.getCharacter().handCards.sort((a, b) => a.type - b.type);
+  cardPickUser.getCharacter().gold -= pickedCardPrice;
 
   // 상점 목록에서 해당 상품 삭제
-  const removedCardIndex = cards.indexOf(fleMarketPickCard);
+  const removedCardIndex = marketItemsLists.indexOf(fleMarketPickCard);
   if (removedCardIndex !== -1) {
-    cards.splice(removedCardIndex, 1);
+    marketItemsLists.splice(removedCardIndex, 1);
   }
-  redisFleaMarketCards[room.id] = cards;
-  await setRedisData('fleaMarketCards', redisFleaMarketCards);
-  await userUpdateNotification(room);
-  await setRedisData('roomData', rooms);
+
+  userUpdateNotification(room);
   sendPacket(socket, config.packetType.FLEA_MARKET_CARD_PICK_RESPONSE, {
-    userId: cardPickUser.id,
-    handCards: cardPickUser.character.handCards
+    userId: cardPickUser.getId(),
+    handCards: cardPickUser.getCharacter().handCards
   });
 };
