+++ src/gameServer/handlers/room/createRoomHandler.ts
@@ -1,136 +1,103 @@
-import net from 'net';
 import { sendPacket } from '../../../packet/createPacket.js';
 import { config } from '../../../config/config.js';
-import { getRedisData, getUserIdBySocket, setRedisData } from '../handlerMethod.js';
-import { CreateRoomPayload, CustomSocket, Room, User } from '../../interface/interface.js';
+import { CreateRoomPayload, CustomSocket, RoomData } from '../../interface/interface.js';
 import { GlobalFailCode, RoomStateType } from '../enumTyps.js';
-import { socketSessions } from '../../session/socketSession.js';
 import Server from '../../class/server.js';
+import GameRoom from '../../class/room.js';
+import { convertSendRoomData, getUserBySocket } from '../handlerMethod.js';
+import UserSessions from '../../class/userSessions.js';
 
 let gRoomId: number = 1;
 export const getgRoomId = () => {
-  return gRoomId;
+    return gRoomId;
 };
 export const addgRoomId = () => {
-  gRoomId++;
+    gRoomId++;
 };
 
-export const createRoomHandler = async (socket: net.Socket, payload: Object) => {
-  let failCode: Number = GlobalFailCode.NONE;
-  let success: boolean = true;
-  let roomDatas: Room[] = [];
-  let newRoom;
+export const createRoomHandler = async (socket: CustomSocket, payload: Object) => {
+    let failCode: Number = GlobalFailCode.NONE;
+    let success: boolean = true;
 
-  const createRoomPayload = payload as CreateRoomPayload;
+    let sendRoomInfo;
 
-  const userId: number | null = await getUserIdBySocket(socket as CustomSocket);
-  if (!userId) {
-    success = false;
-    failCode = GlobalFailCode.CREATE_ROOM_FAILED;
-  }
-  else {
-    // 이미 참여중인 방이 있는지 검사
-    const rooms = await getRedisData('roomData');
-    if (rooms) {
-      for (let i = 0; i < rooms.length; i++) {
-        for (let j = 0; j < rooms[i].users.length; j++) {
-          if (rooms[i].users[j].id === userId) {
-            console.error('이미 참여중인 방이 존재합니다.');
-            const sendData = {
-              success: false,
-              room: {},
-              failCode: GlobalFailCode.JOIN_ROOM_FAILED
-            };
-            sendPacket(socket, config.packetType.JOIN_ROOM_RESPONSE, sendData);
-            return;
-          }
-        }
-      }
-    }
+    const createRoomPayload = payload as CreateRoomPayload;
 
-    // 클라로부터 받은 roomName과 maxUserNum이 undefine과 null이 아닐 경우
-    if (!createRoomPayload.name || !createRoomPayload.maxUserNum) {
-      success = false;
-      failCode = GlobalFailCode.CREATE_ROOM_FAILED;
+    createRoomPayload.maxUserNum = 5;
+
+    const user = getUserBySocket(socket);
+    if (!user) {
+        success = false;
+        failCode = GlobalFailCode.CREATE_ROOM_FAILED;
     }
     else {
-      const users: User[] = [];
-      let user: User | null = null;
-      const userDatas = await getRedisData('userData');
-      if (userDatas) {
-        for (let i = 0; i < userDatas.length; i++) {
-          if (socketSessions[userDatas[i].id] === socket) {
-            user = userDatas[i];
-          }
-        }
-      }
-
-      if (!user) {
-        console.log("createRoomHandler user가 없음");
-        return;
-      }
-
-      users.push(user);
-      createRoomPayload.maxUserNum = 5;
-
-      roomDatas = await getRedisData('roomData');
-      if (!roomDatas) {
-        newRoom = {
-          id: gRoomId,
-          ownerId: userId,
-          ownerEmail: user.email,
-          name: createRoomPayload.name,
-          maxUserNum: createRoomPayload.maxUserNum,
-          state: RoomStateType.WAIT,
-          users: users
-        };
-
-        roomDatas = [newRoom];
-
-        await setRedisData('roomData', roomDatas);
-
-        gRoomId++;
-      } else {
-        let existRoom = false;
-
-        // 방을 이미 만들었는지 검사
-        for (let i = 0; i < roomDatas.length; i++) {
-          if (roomDatas[i].ownerId === userId) {
-            existRoom = true;
-            break;
-          }
+        const rooms = Server.getInstance().getRooms();
+        if (!rooms) {
+            return;
         }
 
-        if (existRoom === false) {
-          newRoom = {
-            id: gRoomId,
-            ownerId: userId,
-            ownerEmail: user.email,
-            name: createRoomPayload.name,
-            maxUserNum: createRoomPayload.maxUserNum,
-            state: RoomStateType.WAIT,
-            users: users
-          };
+        for (let i = 0; i < rooms.length; i++) {
+            for (let j = 0; j < rooms[i].getUsers().length; j++) {
+                if (rooms[i].getUsers()[j].getId() === user.getId()) {
+                    console.error(`이미 참여중인 방 [${rooms[i]}] 이 있습니다.`)
+                    const sendData = {
+                        success: false,
+                        room: {},
+                        failCode: GlobalFailCode.JOIN_ROOM_FAILED
+                    };
 
-          roomDatas.push(newRoom);
-
-          await setRedisData('roomData', roomDatas);
-
-          gRoomId++;
-        } else {
-          success = false;
-          failCode = GlobalFailCode.CREATE_ROOM_FAILED;
+                    sendPacket(socket, config.packetType.JOIN_ROOM_REQUEST, sendData);
+                    return;
+                }
+            }
         }
-      }
 
-      Server.getInstance().chattingServerSend(
-        config.chattingPacketType.CHATTING_CREATE_ROOM_REQUEST, { email: user.email });
+        if (!createRoomPayload.name || !createRoomPayload.maxUserNum) {
+            success = false;
+            failCode = GlobalFailCode.CREATE_ROOM_FAILED;
+        }
+        else {
+            const rooms = Server.getInstance().getRooms();
+            if (!rooms) {
+                return;
+            }
+
+            const room = rooms.find((room: GameRoom) => room.getRoomOwnerId() === user.getId());
+            if (!room) {
+                const users: UserSessions[] = [];
+                users.push(user);
+
+                const newRoomInfo: RoomData = {
+                    id: gRoomId,
+                    ownerId: user.getId(),
+                    ownerEmail: user.getEmail(),
+                    name: createRoomPayload.name,
+                    maxUserNum: createRoomPayload.maxUserNum,
+                    state: RoomStateType.WAIT,
+                    users: users
+                }
+
+                const newGameRoom: GameRoom = new GameRoom(newRoomInfo);
+                newGameRoom.setUsers(users);
+                Server.getInstance().getRooms().push(newGameRoom);
+
+                sendRoomInfo = convertSendRoomData(newGameRoom);
+
+                gRoomId++;
+            }
+            else {
+                success = false;
+                failCode = GlobalFailCode.CREATE_ROOM_FAILED;
+            }
+
+            Server.getInstance().chattingServerSend(
+                config.chattingPacketType.CHATTING_CREATE_ROOM_REQUEST, { email: user.getEmail() });
+        }
+
+        sendPacket(socket, config.packetType.CREATE_ROOM_RESPONSE, {
+            success,
+            room: sendRoomInfo,
+            failCode
+        });
     }
-
-    sendPacket(socket, config.packetType.CREATE_ROOM_RESPONSE, {
-      success,
-      room: newRoom,
-      failCode
-    });
-  }
 };
