+++ src/gameServer/handlers/room/getRoomListHandler.ts
@@ -1,35 +1,41 @@
 import { sendPacket } from '../../../packet/createPacket.js';
 import { config } from '../../../config/config.js';
-import { getRedisData, getUserIdBySocket } from '../handlerMethod.js';
+import { convertSendRoomData, getRoomByUserId, getUserBySocket } from '../handlerMethod.js';
 import { CustomSocket, Room } from '../../interface/interface.js';
 import { GlobalFailCode, RoomStateType } from '../enumTyps.js';
+import Server from '../../class/server.js';
 
 export const getRoomListHandler = async (socket: CustomSocket) => {
-  const rooms: Room[] | undefined = await getRedisData('roomData');
+    const rooms = Server.getInstance().getRooms();
 
-  const userId: number | null = await getUserIdBySocket(socket);
-  if (rooms === undefined) return;
-  for (let i = 0; i < rooms.length; i++) {
-    for (let j = 0; j < rooms[i].users.length; j++) {
-      if (rooms[i].users[j].id === userId && rooms[i].state === RoomStateType.WAIT) {
-        const sendData = {
-          success: true,
-          room: rooms[i],
-          failCode: GlobalFailCode.NONE
-        };
-        sendPacket(socket, config.packetType.JOIN_ROOM_RESPONSE, sendData);
-      }
+    //console.log("방 목록 요청");    
+
+    const user = getUserBySocket(socket);
+    if (!user) {
+        console.log("getRoomListHandler user 없음")
+        return;
     }
-  }
 
-  const waitRooms = [];
-  for (let i = 0; i < rooms.length; i++) {
-    if (rooms[i].state === RoomStateType.WAIT) waitRooms.push(rooms[i]);
-  }
-  if (!rooms) {
-    console.error('getRoomListHandler 방 목록을 찾을 수 없음');
-    return;
-  }
+    const room = getRoomByUserId(user.getId());
+    if (room && room.getRoomState() !== RoomStateType.INGAME) {
+        const sendRoomData = convertSendRoomData(room);
+        const sendData = {
+            success: true,
+            room: sendRoomData,
+            failCode: GlobalFailCode.NONE
+        }
 
-  sendPacket(socket, config.packetType.GET_ROOM_LIST_RESPONSE, { rooms: waitRooms });
+        sendPacket(socket, config.packetType.JOIN_ROOM_RESPONSE, sendData);
+        return;
+    }
+
+    const sendRooms: Room[] = [];
+    for (let i = 0; i < rooms.length; i++) {
+        if (rooms[i].getRoomState() === RoomStateType.WAIT) {
+            const sendRoomData = convertSendRoomData(rooms[i]);
+            sendRooms.push(sendRoomData);
+        }
+    }
+
+    sendPacket(socket, config.packetType.GET_ROOM_LIST_RESPONSE, { rooms: sendRooms });
 };
