+++ src/protobuf/packetNames.ts
@@ -17,7 +17,7 @@
     C2SFleaMarketPickRequest: 'request.C2SFleaMarketPickRequest',
     C2SFleaMarketSellRequest: 'request.C2SFleaMarketSellRequest',
     C2SFleaMarketCardPickRequest: 'request.C2SFleaMarketCardPickRequest',
-    C2SReactionRequest: 'request.C2SReactionRequest',
+    C2SResultRequest: 'request.C2SResultRequest',
     C2SPassDebuffRequest: 'request.C2SPassDebuffRequest',
     C2SDestroyCardRequest: 'request.C2SDestroyCardRequest',
     C2SCardSelectRequest: 'request.C2SCardSelectRequest',
@@ -36,7 +36,7 @@
     S2CUseCardResponse: 'response.S2CUseCardResponse',
     S2CFleaMarketPickResponse: 'response.S2CFleaMarketPickResponse',
     S2CFleaMarketCardPickResponse: 'response.S2CFleaMarketCardPickResponse',
-    S2CReactionResponse: 'response.S2CReactionResponse',
+    S2CBossRoundResponse: 'response.S2CBossRoundResponse',
     S2CDestroyCardResponse: 'response.S2CDestroyCardResponse',
     S2CCardSelectResponse: 'response.S2CCardSelectResponse',
     S2CPassDebuffResponse: 'response.S2CPassDebuffResponse',
